+++ src/types.rs
@@ -1,5 +1,3 @@
 //! Type definitions.
 
-use r2d2_redis::RedisConnectionManager;
-
-pub type RedisPool = r2d2::Pool<RedisConnectionManager>;
+pub type RedisPool = r2d2::Pool<redis::Client>;